Clean up snakefile and main workflow
After updates to the subsampling format of the pipeline roughly half of the process was deprecated and would simply be confusing to anyone trying to view the code.
The unnecessary steps have now been removed for clarity and to save space.
The pipeline now correctly names files to have the `sars-cov-2-belgium` prefix, rather than using the nextstrain default `ncov`--the script previously used to rename files has been moved to a more appropriate directory for clarity.
This commit also removes a deprecated file containing sequence cluster identifications.